@@ -15,23 +15,23 @@
     "logs": "firebase functions:log"
   },
   "dependencies": {
-    "@google-cloud/storage": "^2.3.4",
-    "child-process-promise": "^2.2.1",
-    "firebase-admin": "6.4.0",
-    "firebase-functions": "^2.0.5",
+    "date-fns": "^2.0.0-alpha.26",
+    "deep-equal": "^1.0.1",
+    "firebase-admin": "^6.4.0",
+    "firebase-functions": "^2.1.0",
     "generate-google-calendar-url": "^0.0.2",
-    "ical-generator": "^1.0.3",
-    "moment": "^2.22.2",
+    "handlebars": "^4.0.12",
+    "ical-generator": "^1.4.2",
+    "moment": "^2.23.0",
     "moment-range": "^4.0.1",
-    "nodemailer": "^4.6.8",
-    "qrcode": "^1.2.2",
+    "nodemailer": "^4.7.0",
+    "qrcode": "^1.3.2",
     "sharp": "^0.21.1"
   },
   "devDependencies": {
-    "babel-cli": "^6.26.0",
-    "babel-plugin-transform-object-rest-spread": "^6.26.0",
-    "babel-preset-env": "^1.7.0",
-    "eslint": "^5.4.0",
+    "@babel/cli": "^7.2.3",
+    "@babel/preset-env": "^7.2.3",
+    "eslint": "^5.10.0",
     "eslint-plugin-promise": "^4.0.1"
   },
   "private": true
