@@ -1,7 +1,7 @@
-import React from 'react'
-
+import React from "react"
+import PropTypes from "prop-types"
 import {TableRow, TableCell, Typography} from "@material-ui/core"
-import Loading from './Loading'
+import Loading from "./Loading"
 
 
 const EmptyTableBody = ({title=<Loading isEmpty/>}) =>
@@ -9,12 +9,13 @@ const EmptyTableBody = ({title=<Loading isEmpty/>}) =>
     <TableCell colSpan={10}>
       <Typography
         align="center"
-        variant="subheading"
+        variant="subtitle1"
       >
         {title}
       </Typography>
     </TableCell>
   </TableRow>
 
+EmptyTableBody.propTypes = {title: PropTypes.element}
 
 export default EmptyTableBody
