@@ -1,10 +1,10 @@
 import React, {Component} from "react"
-import {withRouter} from 'react-router-dom'
-import PropTypes from 'prop-types'
+import {withRouter} from "react-router-dom"
+import PropTypes from "prop-types"
 
 
-import {Dialog, Button, DialogActions, DialogTitle, DialogContent} from '@material-ui/core'
-import {withStore} from "../App/Store"
+import {Dialog, Button, DialogActions, DialogTitle, DialogContent} from "@material-ui/core"
+import {withStore} from "../../db"
 
 class Modal extends Component {
 
@@ -76,7 +76,7 @@ class Modal extends Component {
           <Button
             color="primary"
             onClick={this.handleClose}
-            variant="flat"
+            variant="text"
           >
             {cancelLabel || "Mégse"}
           </Button>
