@@ -1,12 +1,13 @@
 import React from "react"
+import PropTypes from "prop-types"
 import {Button} from "@material-ui/core"
-import Add from '@material-ui/icons/ShortTextRounded'
-import {DB} from "../../../utils/firebase"
-import {withStore} from "../../App/Store"
+import Add from "@material-ui/icons/ShortTextRounded"
+import {DB} from "../../../lib/firebase"
+import {withStore} from "../../../db"
 
 
 const NewParagraph = ({
-  sendNotification, path
+  sendNotification, path, relativeFAB
 }) => {
   const handleCreateNewParagraph = () =>
     DB.ref(`paragraphs${path}`)
@@ -20,15 +21,28 @@ const NewParagraph = ({
       }))
       .catch(sendNotification)
 
+  let style
+
+  if (relativeFAB) {
+    style ={
+      position: "relative",
+      left: "calc(100% - 16px)",
+      transform: "translateX(-100%)",
+      bottom: -8
+    }
+  } else {
+    style={
+      position: "fixed",
+      right: 32,
+      bottom: 32
+    }
+  }
+
   return (
     <Button
       color="secondary"
       onClick={handleCreateNewParagraph}
-      style={{
-        position: "fixed",
-        right: 32,
-        bottom: 32
-      }}
+      style={style}
       variant="extendedFab"
     >
       <Add/>
@@ -36,6 +50,12 @@ const NewParagraph = ({
     </Button>
 
   )
+}
+
+NewParagraph.propTypes = {
+  sendNotification: PropTypes.func,
+  path: PropTypes.string,
+  relativeFAB: PropTypes.bool
 }
 
 export default withStore(NewParagraph)
