@@ -1,11 +1,11 @@
-import React, {Component, Fragment} from "react"
-import {PARAGRAPHS_DB} from "../../../utils/firebase"
+import React, {Component} from "react"
+import {PARAGRAPHS_DB} from "../../../lib/firebase"
 
 import {Grid, TextField, CardActions, Button, Divider, Card, CardContent, Typography, CardHeader, Tooltip} from "@material-ui/core"
 import {SortableHandle} from "react-sortable-hoc"
 
 import DragIndicator from "@material-ui/icons/DragIndicator"
-import {withStore} from "../../App/Store"
+import {withStore} from "../../../db"
 
 const DragHandle = SortableHandle(
   () => (
@@ -76,6 +76,10 @@ class TextParagraph extends Component {
   }
 
   render() {
+    if (!this.props.item) {
+      return null
+    }
+
     const {
       value, isEditing
     } = this.state
@@ -120,7 +124,7 @@ class TextParagraph extends Component {
         <CardActions>
           <Grid container justify="flex-end">
             {isEditing ?
-              <Fragment>
+              <>
                 <Button
                   onClick={this.handleCloseEdit}
                   size="small"
@@ -133,8 +137,8 @@ class TextParagraph extends Component {
                 >
                   Mentés
                 </Button>
-              </Fragment> :
-              <Fragment>
+              </> :
+              <>
                 <Button
                   color="secondary"
                   onClick={this.handleDeleteParagraph}
@@ -149,7 +153,7 @@ class TextParagraph extends Component {
                 >
                   Módosít
                 </Button>
-              </Fragment>
+              </>
             }
           </Grid>
         </CardActions>
