@@ -1,17 +1,18 @@
-import Modal from './Modal'
-import Tip from './Tip'
-import Post from './Post'
-import Loading from './Loading'
-import ExtendedFAB from './ExtendedFAB'
-import Paragraphs from './Paragraphs'
-import Background from './Background'
-import Gallery from './Gallery'
-import EmptyTableBody from './EmptyTableBody'
-import withToolbar from './FilterToolbar'
+import Modal from "./Modal"
+import Tip from "./Tip"
+import Post from "./Post"
+import Loading from "./Loading"
+import ExtendedFAB from "./ExtendedFAB"
+import Paragraphs from "./Paragraphs"
+import Background from "./Background"
+import Gallery from "./Gallery"
+import EmptyTableBody from "./EmptyTableBody"
+import withToolbar from "./FilterToolbar"
+import Item from "./Item"
 
 
 export {
   Modal, Post, Tip, Loading, ExtendedFAB,
-  Paragraphs, Background, Gallery, EmptyTableBody,
+  Paragraphs, Background, Gallery, EmptyTableBody, Item,
   withToolbar
 }
