@@ -1,6 +1,8 @@
+import {validateReservation} from "./validate"
+import theme from "./theme"
+import {routes, toRoute} from "./routes"
+import colors from "./colors"
 
-export {isValidReservation} from './validate'
-export {theme} from './theme'
-export {moment} from './moment'
-export {routes, toRoute} from './routes'
-export {colors} from './colors'
+export {
+  colors, routes, toRoute, theme, validateReservation
+}
