@@ -8,7 +8,7 @@ const routes = {
   FACEBOOK: "https://www.facebook.com/B%C3%ADbic-Vendegh%C3%A1zak-%C3%89s-S%C3%B6r%C3%B6z%C5%91-200199203718517",
   LOGIN: "/bejelentkezes",
   WELCOME: "/kezdolap",
-  SPECIAL_OFFER: "/ajanlatok",
+  MESSAGES: "/uzenetek",
   ROOMS: "/szobak",
   INTRO: "/bemutatkozas",
   CERTIFICATES: "/tanusitvanyok",
@@ -20,6 +20,7 @@ const routes = {
   RESERVATIONS: "/foglalasok",
   FEEDBACKS: "/visszajelzesek",
   STATS: "/statisztikak",
+  GALLERY: "galeria",
   EDIT: "szerkeszt",
   BLOCK: "blokkolas",
   NEW: "uj",
@@ -32,7 +33,7 @@ const routes = {
 
 /**
  * Construct a route path from the parameters
- * @param  {...string} parts individual parts of the desired path
+ * @param  {...string} individual parts of the desired path
  * @returns {string} a string of parts joined by a /
  */
 const toRoute = (...parts) => [...parts].join("/")
