@@ -1,9 +1,6 @@
-import {createMuiTheme} from '@material-ui/core/styles'
-import {colors} from './colors'
+import {createMuiTheme} from "@material-ui/core/styles"
+import colors from "./colors"
 
-const theme = createMuiTheme({palette: {
-  primary: {main: colors.lightBrown},
-  secondary: {main: colors.orange}
-}})
-
-export {theme}
+export default createMuiTheme({palette: {primary: {main: colors.lightBrown},
+  secondary: {main: colors.orange}},
+typography: {useNextVariants: true}})
